Move confetti component to the dashboard/components path
