+++ client/dashboard/components/confetti.ts
